requirements.txt
# Core dependencies for Phase 1 (Python 3.12 compatible)
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
networkx>=3.1
matplotlib>=3.8.0
seaborn>=0.13.0
yfinance>=0.2.28
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
jupyter>=1.0.0
plotly>=5.17.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0